requirements.txt
aniso8601==9.0.1
apispec==5.1.1
astroid==2.11.2
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
dill==0.3.4
Flask==2.1.1
idna==3.3
importlib-metadata==4.11.3
iniconfig==1.1.1
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.1
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
marshmallow==3.15.0
mccabe==0.7.0
packaging==21.3
platformdirs==2.5.1
pluggy==1.0.0
py==1.11.0
pydantic==1.9.0
pylint==2.13.5
pyparsing==3.0.7
pytest==7.1.1
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
six==1.16.0
tomli==2.0.1
typing-extensions==4.1.1
urllib3==1.26.9
webargs==8.1.0
Werkzeug==2.1.1
wrapt==1.14.0
zipp==3.8.0
